elevennote/requirements/dev.txt
-r base.txt
django-debug-toolbar==1.11
coverage==4.5.3
